+++ components/Dialog/Dialog.tsx
@@ -1,0 +1,108 @@
+"use client";
+
+import type { ReactNode } from "react";
+import { useRef, useEffect } from "react";
+
+import STYLES from "./Dialog.module.scss";
+
+type Props = {
+  id: string;
+  children: ReactNode;
+  handleResetFilters: () => void;
+  isOpen: boolean;
+  onClose: () => void | null;
+  totalFilteredTunnellers: number;
+  totalTunnellers: number;
+};
+
+type DialogProps = {
+  isDialogOpen: boolean;
+};
+
+// TODO: this check if the browser support the HTML dialog element. We can remove it once we drop support as a business for Safari 14
+const dialogSupported = typeof HTMLDialogElement === "function";
+
+const setPageProperties = ({ isDialogOpen }: DialogProps) => {
+  document.body.style.overflowY = isDialogOpen ? "hidden" : "visible";
+
+  if (!dialogSupported) {
+    document.body.style.position = isDialogOpen ? "fixed" : "relative";
+    document.body.style.width = isDialogOpen ? "100%" : "auto";
+  }
+};
+
+export function Dialog(props: Props) {
+  const {
+    children,
+    id,
+    isOpen,
+    handleResetFilters,
+    onClose,
+    totalFilteredTunnellers,
+    totalTunnellers,
+  } = props;
+
+  const ref = useRef<HTMLDialogElement>(null);
+
+  useEffect(() => {
+    if (isOpen) {
+      // There is a bug on older versions of browser using chromium (chrome, firefox, edge >114) where the dialog got an open attribute even before it is opened.
+      // Therefore, when trying to open it, it crashes and log an error mentioning the dialog has already an open attribute.
+      ref.current?.removeAttribute("open");
+      ref.current?.showModal?.();
+    } else {
+      ref.current?.close?.();
+    }
+
+    setPageProperties({ isDialogOpen: isOpen });
+    return () => {
+      setPageProperties({ isDialogOpen: false });
+    };
+  }, [isOpen]);
+
+  return (
+    <>
+      {isOpen ? (
+        <div
+          className={
+            STYLES[`dialog-wrapper ${dialogSupported ? "" : "dialog-polyfill"}`]
+          }
+        >
+          {!dialogSupported && (
+            <div id={`${id}-polyfill"`} data-open={isOpen} />
+          )}
+          <dialog
+            id={id}
+            className={STYLES.dialog}
+            onClose={onClose}
+            data-open={isOpen}
+            ref={ref}
+          >
+            <div className={STYLES["dialog-header-container"]}>
+              <div className={STYLES["dialog-header"]}>
+                <h2>Filter</h2>
+                <button onClick={onClose} className={STYLES["close-button"]}>
+                  Done
+                </button>
+              </div>
+            </div>
+            <div className={STYLES["dialog-container"]}>{children}</div>
+            <div className={STYLES["dialog-footer-container"]}>
+              <div className={STYLES["dialog-footer"]}>
+                <button
+                  className={STYLES["dialog-cancel-button"]}
+                  onClick={handleResetFilters}
+                >
+                  Remove filter
+                </button>
+                <div className={STYLES["total-filters"]}>
+                  {totalFilteredTunnellers}/{totalTunnellers}
+                </div>
+              </div>
+            </div>
+          </dialog>
+        </div>
+      ) : null}
+    </>
+  );
+}
